Delegate most methods, and provide a Drawable.
